+++ crates/ruma-common/src/events/call/candidates.rs
@@ -2,16 +2,11 @@
 //!
 //! [`m.call.candidates`]: https://spec.matrix.org/v1.2/client-server-api/#mcallcandidates
 
-#[cfg(feature = "unstable-msc2746")]
-use js_int::uint;
 use js_int::UInt;
 use ruma_macros::EventContent;
 use serde::{Deserialize, Serialize};
 
-#[cfg(feature = "unstable-msc2746")]
-use super::CallVersion;
-#[cfg(feature = "unstable-msc2746")]
-use crate::OwnedVoipId;
+use crate::{OwnedVoipId, VoipVersionId};
 
 /// The content of an `m.call.candidates` event.
 ///
@@ -21,20 +16,12 @@ use crate::OwnedVoipId;
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
 #[ruma_event(type = "m.call.candidates", kind = MessageLike)]
 pub struct CallCandidatesEventContent {
-    #[cfg(not(feature = "unstable-msc2746"))]
     /// A unique identifier for the call.
-    ///
-    /// With the `unstable-msc2746` feature, this uses the stricter `VoipId` type.
-    pub call_id: String,
-
-    #[cfg(feature = "unstable-msc2746")]
-    /// A unique identifier for the call.
-    ///
-    /// Without the `unstable-msc2746` feature, this can be any string.
     pub call_id: OwnedVoipId,
 
     #[cfg(feature = "unstable-msc2746")]
-    /// **Required in version 1.** The unique ID for this session for the duration of the call.
+    /// **Required in VoIP version 1.** The unique ID for this session for the duration of the
+    /// call.
     ///
     /// Must be the same as the one sent by the previous invite or answer from
     /// this session.
@@ -43,60 +30,36 @@ pub struct CallCandidatesEventContent {
 
     /// A list of candidates.
     ///
-    /// With the `unstable-msc2746` feature, in version 1, this list should end with a
+    /// With the `unstable-msc2746` feature, in VoIP version 1, this list should end with a
     /// `Candidate` with an empty `candidate` field when no more candidates will be sent.
     pub candidates: Vec<Candidate>,
 
-    #[cfg(not(feature = "unstable-msc2746"))]
     /// The version of the VoIP specification this messages adheres to.
-    ///
-    /// With the `unstable-msc2746` feature, this can be a `UInt` or a `String`.
-    pub version: UInt,
-
-    #[cfg(feature = "unstable-msc2746")]
-    /// The version of the VoIP specification this messages adheres to.
-    ///
-    /// Without the `unstable-msc2746` feature, this is a `UInt`.
-    pub version: CallVersion,
+    pub version: VoipVersionId,
 }
 
 impl CallCandidatesEventContent {
     /// Creates a new `CallCandidatesEventContent` with the given call id, candidate list and VoIP
     /// version.
-    ///
-    /// With the `unstable-msc2746` feature, this method takes an `OwnedVoipId` for the call ID and
-    /// a `CallVersion` for the version.
-    #[cfg(not(feature = "unstable-msc2746"))]
-    pub fn new(call_id: String, candidates: Vec<Candidate>, version: UInt) -> Self {
-        Self { call_id, candidates, version }
-    }
-
-    /// Creates a new `CallCandidatesEventContent` with the given call id, candidate list and VoIP
-    /// version.
-    ///
-    /// Without the `unstable-msc2746` feature, this method takes a `String` for the call ID and a
-    /// `UInt` for the version.
-    #[cfg(feature = "unstable-msc2746")]
-    pub fn new(call_id: OwnedVoipId, candidates: Vec<Candidate>, version: CallVersion) -> Self {
+    pub fn new(call_id: OwnedVoipId, candidates: Vec<Candidate>, version: VoipVersionId) -> Self {
         Self { call_id, candidates, version, party_id: None }
     }
 
-    /// Convenience method to create a version 0 `CallCandidatesEventContent` with all the required
-    /// fields.
-    #[cfg(feature = "unstable-msc2746")]
+    /// Convenience method to create a VoIP version 0 `CallCandidatesEventContent` with all the
+    /// required fields.
     pub fn version_0(call_id: OwnedVoipId, candidates: Vec<Candidate>) -> Self {
-        Self::new(call_id, candidates, uint!(0).into())
+        Self::new(call_id, candidates, VoipVersionId::V0)
     }
 
-    /// Convenience method to create a version 1 `CallCandidatesEventContent` with all the required
-    /// fields.
+    /// Convenience method to create a VoIP version 1 `CallCandidatesEventContent` with all the
+    /// required fields.
     #[cfg(feature = "unstable-msc2746")]
     pub fn version_1(
         call_id: OwnedVoipId,
         party_id: OwnedVoipId,
         candidates: Vec<Candidate>,
     ) -> Self {
-        Self { call_id, party_id: Some(party_id), candidates, version: uint!(1).into() }
+        Self { call_id, party_id: Some(party_id), candidates, version: VoipVersionId::V1 }
     }
 }