+++ crates/ruma-common/src/events/call/invite.rs
@@ -2,17 +2,16 @@
 //!
 //! [`m.call.invite`]: https://spec.matrix.org/v1.2/client-server-api/#mcallinvite
 
-#[cfg(feature = "unstable-msc2746")]
-use js_int::uint;
 use js_int::UInt;
 use ruma_macros::EventContent;
 use serde::{Deserialize, Serialize};
 
+#[cfg(feature = "unstable-msc2746")]
+use super::CallCapabilities;
 use super::OfferSessionDescription;
 #[cfg(feature = "unstable-msc2746")]
-use super::{CallCapabilities, CallVersion};
-#[cfg(feature = "unstable-msc2746")]
-use crate::{OwnedUserId, OwnedVoipId};
+use crate::OwnedUserId;
+use crate::{OwnedVoipId, VoipVersionId};
 
 /// The content of an `m.call.invite` event.
 ///
@@ -21,20 +20,11 @@ use crate::{OwnedUserId, OwnedVoipId};
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
 #[ruma_event(type = "m.call.invite", kind = MessageLike)]
 pub struct CallInviteEventContent {
-    #[cfg(not(feature = "unstable-msc2746"))]
     /// A unique identifier for the call.
-    ///
-    /// With the `unstable-msc2746` feature, this uses the stricter `OwnedVoipId` type.
-    pub call_id: String,
-
-    #[cfg(feature = "unstable-msc2746")]
-    /// A unique identifier for the call.
-    ///
-    /// Without the `unstable-msc2746` feature, this can be any string.
     pub call_id: OwnedVoipId,
 
     #[cfg(feature = "unstable-msc2746")]
-    /// **Required in version 1.** A unique ID for this session for the duration of the call.
+    /// **Required in VoIP version 1.** A unique ID for this session for the duration of the call.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub party_id: Option<OwnedVoipId>,
 
@@ -47,25 +37,16 @@ pub struct CallInviteEventContent {
     /// The session description object.
     pub offer: OfferSessionDescription,
 
-    #[cfg(not(feature = "unstable-msc2746"))]
     /// The version of the VoIP specification this messages adheres to.
-    ///
-    /// With the `unstable-msc2746` feature, this can be a `UInt` or a `String`.
-    pub version: UInt,
+    pub version: VoipVersionId,
 
     #[cfg(feature = "unstable-msc2746")]
-    /// The version of the VoIP specification this messages adheres to.
-    ///
-    /// Without the `unstable-msc2746` feature, this is a `UInt`.
-    pub version: CallVersion,
-
-    #[cfg(feature = "unstable-msc2746")]
-    /// **Added in version 1.** The VoIP capabilities of the client.
+    /// **Added in VoIP version 1.** The VoIP capabilities of the client.
     #[serde(default, skip_serializing_if = "CallCapabilities::is_default")]
     pub capabilities: CallCapabilities,
 
     #[cfg(feature = "unstable-msc2746")]
-    /// **Added in version 1.** The intended target of the invite, if any.
+    /// **Added in VoIP version 1.** The intended target of the invite, if any.
     ///
     /// If this is `None`, the invite is intended for any member of the room, except the sender.
     ///
@@ -77,30 +58,11 @@ pub struct CallInviteEventContent {
 impl CallInviteEventContent {
     /// Creates a new `CallInviteEventContent` with the given call ID, lifetime, offer and VoIP
     /// version.
-    ///
-    /// With the `unstable-msc2746` feature, this method takes an `OwnedVoipId` for the call ID and
-    /// a `CallVersion` for the version.
-    #[cfg(not(feature = "unstable-msc2746"))]
-    pub fn new(
-        call_id: String,
-        lifetime: UInt,
-        offer: OfferSessionDescription,
-        version: UInt,
-    ) -> Self {
-        Self { call_id, lifetime, offer, version }
-    }
-
-    /// Creates a new `CallInviteEventContent` with the given call ID, lifetime, offer and VoIP
-    /// version.
-    ///
-    /// Without the `unstable-msc2746` feature, this method takes a `String` for the call ID and a
-    /// `UInt` for the version.
-    #[cfg(feature = "unstable-msc2746")]
     pub fn new(
         call_id: OwnedVoipId,
         lifetime: UInt,
         offer: OfferSessionDescription,
-        version: CallVersion,
+        version: VoipVersionId,
     ) -> Self {
         Self {
             call_id,
@@ -115,9 +77,8 @@ impl CallInviteEventContent {
 
     /// Convenience method to create a version 0 `CallInviteEventContent` with all the required
     /// fields.
-    #[cfg(feature = "unstable-msc2746")]
     pub fn version_0(call_id: OwnedVoipId, lifetime: UInt, offer: OfferSessionDescription) -> Self {
-        Self::new(call_id, lifetime, offer, uint!(0).into())
+        Self::new(call_id, lifetime, offer, VoipVersionId::V0)
     }
 
     /// Convenience method to create a version 1 `CallInviteEventContent` with all the required
@@ -135,7 +96,7 @@ impl CallInviteEventContent {
             party_id: Some(party_id),
             lifetime,
             offer,
-            version: uint!(1).into(),
+            version: VoipVersionId::V1,
             capabilities,
             invitee: None,
         }